+++ CHANGELOG.md
@@ -1,6 +1,27 @@
 Release Notes
 =============
 
+## 0.79.1
+
+Released August 10, 2016
+
+**Fixed**
+
+- [ColumnDefinition](https://github.com/groue/GRDB.swift#database-schema) `check` and `references` methods can now define several constraints:
+    
+    ```swift
+    try db.create(table: "users") { t in
+        t.column("name", .Text).notNull()
+            .check { length($0) > 0 }
+            .check { !["root", "admin"].contains($0) }
+    }
+    ```
+
+- [Persistable](https://github.com/groue/GRDB.swift#persistable-protocol) `update`, `exists` and `delete` methods now work with objects that have a nil primary key. They used to crash.
+
+- The `update(_:columns:)` method, which performs partial updates, no longer ignores unknown columns.
+
+
 ## 0.79.0
 
 Released August 8, 2016